refactor : 예외 판별 함수들 utils로 이동
