+++ src/index.js
@@ -1,4 +1,5 @@
 import { text } from './fixtrue';
+import { isNumber, isNot3Digit, isInZero, isInDuplicateDigit } from './utils';
 
 // BaseballGame class
 export default class BaseballGame {
@@ -126,32 +127,6 @@ export class BaseballGameView {
 }
 
 // user input judge function and reset input function
-
-function isNumber(userInputNumbers) {
-  const regex = /^[0-9]*$/;
-  const response = regex.test(userInputNumbers);
-
-  return response;
-}
-
-function isNot3Digit(userInputNumbers) {
-  if (userInputNumbers.length !== 3) {
-    return true;
-  }
-}
-
-function isInZero(userInputNumbers) {
-  if (userInputNumbers.includes('0')) {
-    return true;
-  }
-}
-
-function isInDuplicateDigit(userInputNumbers) {
-  const deduplicateCount = Array.from(new Set(userInputNumbers)).length;
-  if (deduplicateCount !== 3) {
-    return true;
-  }
-}
 
 function resetInputNumbers() {
   const userInput = document.getElementById('user-input');
